UserStory model added to Projects
I created a UserStory model with all other MVP stuff. You cann add a userstory in a project, can add multiple acceptance criteria, persona's (not yet per project), mvp or not and project.
You can view the user stories in ViewProject with separated stories and information.
To things that still need to be done:
* userstories in ViewProject are not yet collapsible because collapsible is not a possibility on a RepeatableEntry
* the function to select only the personas from the specific project does not work yet. Until it works you can select all personas